Move files sample to improve an other lib integration
