+++ tests/ZendTest/Test/PHPUnit/Controller/ModuleDependenciesTest.php
@@ -23,7 +23,7 @@ class ModuleDependenciesTest extends AbstractHttpControllerTestCase
     public function testDependenciesModules()
     {
         $this->setApplicationConfig(
-            include __DIR__ . '/_files/application.config.with.dependencies.php'
+            include __DIR__ . '/../../_files/application.config.with.dependencies.php'
         );
         $sm = $this->getApplicationServiceLocator();
         $this->assertEquals(true, $sm->has('FooObject'));
@@ -37,7 +37,7 @@ class ModuleDependenciesTest extends AbstractHttpControllerTestCase
     public function testBadDependenciesModules()
     {
         $this->setApplicationConfig(
-            include __DIR__ . '/_files/application.config.with.dependencies.disabled.php'
+            include __DIR__ . '/../../_files/application.config.with.dependencies.disabled.php'
         );
         $sm = $this->getApplicationServiceLocator();
         $this->assertEquals(false, $sm->has('FooObject'));
