+++ tests/ZendTest/Test/_files/Baz/src/Baz/Controller/IndexController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mock\Controller;
+namespace Baz\Controller;
 
 use Zend\Mvc\Controller\AbstractActionController;
