+++ tests/ZendTest/Test/_files/application.config.php
@@ -1,12 +1,12 @@
 <?php
 return array(
     'modules' => array(
-        'Mock',
+        'Baz',
     ),
     'module_listener_options' => array(
         'config_static_paths'    => array(),
         'module_paths' => array(
-            'Mock' => __DIR__ . '/Mock/',
+            'Baz' => __DIR__ . '/Baz/',
         ),
     ),
 );
