+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Bar/Module.php
