+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Bar/config/module.config.php
