+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Bar/src/Bar/Controller/IndexController.php
