+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Bar/view/layout/layout.phtml
