+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Foo/Module.php
