+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Foo/config/module.config.php
