+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Foo/src/Foo/Controller/IndexController.php
