+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Foo/view/layout/layout.phtml
