+++ tests/ZendTest/Test/_files/modules-path/with-subdir/Foo/view/mock/index/unittests.phtml
